+++ monitor/asm-arm_inline.h
@@ -1,0 +1,1 @@
+#define isb() asm volatile("isb" : : : "memory")
