+++ monitor/scheduler.h
@@ -7,11 +7,8 @@
 #include "timer.h"
 hvmm_status_t scheduler_init(void);
 
-/* Test switch bmguest. Timer ISR called this function.  */
-void timer_switch_next_guest(void *pdata);
-
 hvmm_status_t scheduler_next_event(int irq, void *pdata);
 
 /* Test Code */
-void timer_test_scheduling(void);
+void scheduler_test_scheduling(void);
 #endif
