+++ monitor/timer.c
@@ -6,72 +6,36 @@
  */
 #include "timer.h"
 #include "generic_timer.h"
-#include "mathlib.h"
 
-//TODO modify list
-//struct timer_source timer_sources;
 static struct timer_channel _channels[TIMER_NUM_MAX_CHANNELS];
 
-/*
- * Called by the hardware timer every 'interval' period
- * Does the hardware timer specific interrupt disable/enable (if necessary)
- */
 static void _timer_hw_callback(void *pdata)
 {
-	/* Generic Timer: disable interrupt */
 	generic_timer_disable_int(GENERIC_TIMER_HYP);
 
-	/* Callback */
-	/* TODO:
-	 * Invoke callbacks in the channel table selectively and conditionally
-	 * i.e. depending on channel's interval value
-	 */
 	_channels[timer_sched].callback(pdata);
 
-    generic_timer_set_tval(GENERIC_TIMER_HYP, _channels[timer_sched].interval);
+    generic_timer_set_tval(GENERIC_TIMER_HYP, _channels[timer_sched].interval_us);
 	
-	/* Generic Timer: enable interrupt */
 	generic_timer_enable_int(GENERIC_TIMER_HYP);
-}
-
-hvmm_status_t timer_register(struct timer_source* ts)
-{
-/*	
-	timer_sources.name = ts->name;
-	timer_sources.start = ts->start;
-	timer_sources.stop = ts->stop;
-	timer_sources.add_callback = ts->add_callback;
-	timer_sources.remove_callback = ts->remove_callback;
-*/
-	return HVMM_STATUS_SUCCESS;
 }
 
 hvmm_status_t timer_init(timer_channel_t channel)
 {	
-	// TODO need such add generic_timer element. before call this function.
-	generic_timer_init(&_channels[channel].ts);
+	generic_timer_init();
 	return HVMM_STATUS_SUCCESS;
 }
 	
 hvmm_status_t timer_start(timer_channel_t channel)
 {
-/*
-	//timer_sources.start(timer_sources.interval);
-	channel[channel].ts.start(channel[channel].interval);
-*/
-
 	if ( _channels[channel].callback != 0 ) {
-    	/* Generic Timer: pass _timer_hw_callback */
+
     	generic_timer_set_callback(GENERIC_TIMER_HYP, &_timer_hw_callback );
     
-    	/* Generic Timer: set interval in 'count' */
-    	/* interval_ms -> count conversion */
-    	generic_timer_set_tval(GENERIC_TIMER_HYP, _channels[channel].interval);
+    	generic_timer_set_tval(GENERIC_TIMER_HYP, _channels[channel].interval_us);
     
-		/* TODO: Firgure out how to avoid making redundant calls more than once */
     	generic_timer_enable_irq(GENERIC_TIMER_HYP);
 
-    	/* Generic Timer: enable interupt */
     	generic_timer_enable_int(GENERIC_TIMER_HYP);
 	}
 
@@ -80,71 +44,34 @@ hvmm_status_t timer_start(timer_channel_t channel)
 
 hvmm_status_t timer_stop(timer_channel_t channel)
 {
-/*
-	channel[channel].ts.stop();
-*/
-
-	/* Generic Timer: disable interupt */
 	generic_timer_disable_int(GENERIC_TIMER_HYP);
 	return HVMM_STATUS_SUCCESS;
 }
 
-hvmm_status_t timer_set_interval(timer_channel_t channel, uint32_t interval)
+hvmm_status_t timer_set_interval(timer_channel_t channel, uint32_t interval_us)
 {
-	_channels[channel].interval = interval;
+	_channels[channel].interval_us = timer_t2c(interval_us);
 	return HVMM_STATUS_SUCCESS;
 }
 
-hvmm_status_t timer_get_interval(timer_channel_t channel, uint32_t *interval)
+uint32_t timer_get_interval(timer_channel_t channel)
 {
-	*interval = _channels[channel].interval;
-	return HVMM_STATUS_SUCCESS;
+	return _channels[channel].interval_us;
 }
-
 
 hvmm_status_t timer_add_callback(timer_channel_t channel, timer_callback_t callback)
 {
-/*
-	_channels[channel].ts.add_callback(channel, callback);
-*/
-
-	/* TODO:
-	 * Support adding multiple callbacks for a given channel
-	 */
 	_channels[channel].callback = callback;
-
 	return HVMM_STATUS_SUCCESS;
 }
 
 hvmm_status_t timer_remove_callback(timer_channel_t channel, timer_callback_t callback)
 {
-/*
-	channel[channel].ts.remove_callback(0, 0);
-*/
-
-	/* TODO:
-	 * Implement removing THE callback from the list of the channel
-	 */
 	_channels[channel].callback = 0;
-
 	return HVMM_STATUS_SUCCESS;
 }
 
-void timer_get_time(struct timeval* timeval)
+uint64_t timer_t2c(uint64_t time_us)
 {
-	uint64_t cnt = read_cntpct();
-	timeval->tv_sec = udiv64(timer_c2t(cnt), 1000);
-	timeval->tv_usec = timer_c2t(cnt) *  1000;
-/*
-	_channels[0].ts.get_time(timeval);
-*/
-}
-
-
-uint64_t timer_c2t(uint64_t count){
-	return udiv64(count, COUNT_PER_MSEC);
-} 
-
-uint64_t timer_t2c(uint64_t time){
-	return time * COUNT_PER_MSEC;
+	return time_us * COUNT_PER_USEC;
 }
