+++ monitor/timer.h
@@ -4,157 +4,96 @@
 #include "hvmm_types.h"
 #include "arch_types.h"
 
-#if defined(ARCH_FAST_MODEL)
-    #define COUNT_PER_SEC 100000000
-    #define COUNT_PER_MSEC 100000
-    #define COUNT_PER_USEC 100
-#else
-    #define COUNT_PER_SEC 100000000
-    #define COUNT_PER_MSEC 100000
+#if defined(BOARD_RTSMVE_CA15)
     #define COUNT_PER_USEC 100
 #endif
 
-/**
- *  Implements Timer functionality such as,
+/*
+ * Implements Timer functionality such as,
  *
- *	- timer channel: periodic callback at a given time interval
- *  - current time: time since boot-up
+ * - timer channel: periodic callback at a given time interval
+ * - current time: time since boot-up
  *
- *	==Example Usage==
+ * ==Example Usage==
  *
- *	Initialize the timer module
- *		timer_init(timer_sched);
+ * Initialize the timer module
+ *  timer_init(timer_sched);
  *
- *	Set interval 10 to timer_sched channel
- *		timer_set_interval( timer_sched, 10 );
+ * Set interval 10 to timer_sched channel
+ *  timer_set_interval( timer_sched, 10 );
  *
- *	Set cb1 callback function to timer_scehd channel.
- *	cb1 call after 10 interval.
- *		timer_set_callback( timer_sched, &cb1 );
+ * Set cb1 callback function to timer_scehd channel.
+ * cb1 call after 10 interval.
+ *  timer_set_callback( timer_sched, &cb1 );
  *
- *	Start timer_sched channel. cb1 call after 10 interval.
- *		timer_start( timer_sched );
- *
- *		...
- *
- *	Remove the call back
- *		timer_remove_callback(timer_sched, &cb1);
-*/
+ * Start timer_sched channel. cb1 call after 10 interval.
+ *  timer_start( timer_sched );
+ */
 
 
 /*
-*	Timer channel
-*	Have own's timer interrupt handler.
-*/
+ * Timer channel
+ * Have own's timer interrupt handler.
+ */
 typedef enum{
     timer_sched = 0,
     TIMER_NUM_MAX_CHANNELS
 } timer_channel_t;
 
 /*
-*   tv_sec : seconds
-*   tv_usec : microseconds
-*/
-struct timeval{
-    uint64_t tv_sec;
-    uint64_t tv_usec;
-};
-
-/*
-*	timer_callback_t
-*/
+ * Timer_callback_t
+ */
 typedef void(*timer_callback_t)(void *pdata);
 
-///////////////////////////////////
-//TODO Need organization this struct.
-/*
-*	Timer device list sturct.
-*/
-
-struct timer_source{
-	const char *name;
-//	hvmm_status_t (*init) (void);
-	hvmm_status_t (*start) (uint32_t interval);
-	hvmm_status_t (*stop) (void);
-	hvmm_status_t (*add_callback)(int, timer_callback_t);
-	hvmm_status_t (*remove_callback)(int, timer_callback_t);
-//	uint32_t interval;
-};
-	
 struct timer_channel{
-	struct timer_source ts;
-	uint32_t interval;
+	uint32_t interval_us;
 	timer_callback_t callback;
 };
 
 /*
- * TODO Need move other source, e.g. module_init.c, list.c
- */
-struct timer_source_list{
-	struct timer_source source;
-	struct timer_source_list *next;
-	struct timer_source_list *prev;
-};
-struct dlist{
-	struct dlist *next, *prev;
-};
-
-///////////////////////////////////
-
-/**
  * Calling this function is required only once in the entire system prior to calls 
  * to other functions of Timer module.
  */
 	hvmm_status_t timer_init(timer_channel_t channel);
 
-/**
+/*
  * Starts the timer channel specified by 'channel'. The callback, 
  * if set, will be periodically called until it's unset or the channel stops by 
  * 'timer_stop(timer_channel)'
  */
 	hvmm_status_t timer_start(timer_channel_t channel);
 
-/**
+/*
  *	Stops the timer channel specified by 'channel' 
  */
 	hvmm_status_t timer_stop(timer_channel_t channel);
 
-/**
- * Sets time interval, in milliseconds, for the timer channel. 
+/*
+ * Sets time interval, in microseconds, for the timer channel. 
  * If the channel has been started and a callback function is set, it will be called 
  * in the next interval
  */
-	hvmm_status_t timer_set_interval(timer_channel_t channel, uint32_t interval);
+	hvmm_status_t timer_set_interval(timer_channel_t channel, uint32_t interval_us);
 
-/**
+/*
  * Returns the time interval for the timer channel if it was set previously. 
  * Unknown value is returned otherwise.
  */
-	hvmm_status_t timer_get_interval(timer_channel_t channel, uint32_t *interval);
+	uint32_t timer_get_interval(timer_channel_t channel);
 
-/**
+/*
  * Adds a callback function for the timer channel.
  */
 	hvmm_status_t timer_add_callback(timer_channel_t channel, timer_callback_t handler);
-/**
+/*
  * Removes the callback function from the timer channel's registered callback function list 
  * if it previously has been added.
  */
 	hvmm_status_t timer_remove_callback(timer_channel_t channel, timer_callback_t handler);
-/**
- * Returns current time since boot-up.
+
+/* 
+ * Converts from microseconds to system counter. 
  */
-void timer_get_time(struct timeval* timeval);
-
-/**
- * Register timer source.
- */
-hvmm_status_t timer_register(struct timer_source* ts);
-
-/* Converts from system counter to milliseconds. */
-uint64_t timer_c2t(uint64_t count);
-
-/* Converts from milliseconds to system counter. */   
-uint64_t timer_t2c(uint64_t time);
+	uint64_t timer_t2c(uint64_t time_us);
    
 #endif
